// src/app/models/country.model.ts
export class Country {
    name: String;
    topLevelDomain: Array<String>;
    alpha2Code: String;
    alpha3Code: String;
    callingCodes: Array<String>;
    capital: String;
    altSpellings: String;
    region: String;
    subregion: String;
    population: Number;
    latlng: Array<Number>;
    demonym: String;
    area: Number;
    gini: Number;
    timezones: Array<String>;
    borders: Array<String>;
    nativeName: String;
    numericCode: String;
    currencies: Array<Object>;
    languages: Array<Object>;
    translations: Object;
    flag: String;
    regionalBlocs: Array<Object>;
    cioc: String;
}